clase_33/clase_23031/vino_app/viewsets.py
from rest_framework.viewsets import ModelViewSet
from .models import Vino
from .serializer import VinoSerializer

class VinoViewSet(ModelViewSet):
    queryset = Vino.objects.all()
    serializer_class = VinoSerializer